Move accountsDB into database folder
