+++ ironfish/src/wallet/account.ts
@@ -7,8 +7,8 @@ import { Assert } from '../assert'
 import { Transaction } from '../primitives'
 import { Note } from '../primitives/note'
 import { IDatabaseTransaction } from '../storage'
-import { AccountsDB } from './accountsdb'
 import { AccountsValue } from './database/accounts'
+import { AccountsDB } from './database/accountsdb'
 import { DecryptedNotesValue } from './database/decryptedNotes'
 import { SyncTransactionParams } from './wallet'
