+++ ironfish/src/wallet/index.ts
@@ -5,4 +5,4 @@ export * from './account'
 export * from './wallet'
 export { AccountsValue } from './database/accounts'
 export * from './validator'
-export * from './accountsdb'
+export * from './database/accountsdb'
